Extracted duplicate code to DateBase
