+++ backend/module/eCampLib/src/Types/DateTimeUtc.php
@@ -6,27 +6,27 @@ use DateTime;
 use DateTimeZone;
 use JsonSerializable;
 
-class DateTimeUtc extends DateTime implements JsonSerializable {
+class DateTimeUtc extends DateBase implements JsonSerializable {
     protected string $FORMAT = 'Y-m-d\TH:iP';
 
     /**
      * DateTimeUTC constructor.
      *
-     * @param null|string $time
+     * @param null|string       $time     Date in string representation
+     * @param null|DateTimeZone $timezone Default timezone is UTC
+     * @param null|string       $format   Allowed format of the datetime
      *
-     * @throws InvalidFormatException
+     * @throws InvalidDateFormatException
+     * @throws InvalidZoneOffsetException
      */
     public function __construct($time = 'now', DateTimeZone $timezone = null, string $format = null) {
         if (null === $timezone) {
             $timezone = new DateTimeZone('UTC');
+        } elseif ('UTC' !== $timezone->getName()) {
+            throw new InvalidZoneOffsetException('Invalid zone offset: '.$timezone->getName().'. Should be 0');
         }
         if ('now' !== $time) {
-            $format = $format ?? $this->FORMAT;
-
-            $parsedDate = (object) date_parse_from_format($format, $time);
-            if ($parsedDate->error_count > 0) {
-                throw new InvalidDateFormatException('Invalid date format: '.$time.'. Should be '.$format);
-            }
+            $this->checkFormat($time, $format);
         }
         parent::__construct($time, $timezone);
     }
