+++ backend/module/eCampLib/src/Types/DateUtc.php
@@ -2,43 +2,31 @@
 
 namespace eCamp\Lib\Types;
 
-use DateTime;
 use DateTimeZone;
 use JsonSerializable;
 
-class DateUtc extends DateTime implements JsonSerializable {
+class DateUtc extends DateBase implements JsonSerializable {
     protected string $FORMAT = 'Y-m-d';
 
     /**
      * DateTimeUTC constructor.
      *
-     * @param null|string $time
+     * @param null|string       $time     Date in string representation
+     * @param null|DateTimeZone $timezone Default timezone is UTC
+     * @param null|string       $format   Allowed format of the date
      *
      * @throws InvalidDateFormatException
+     * @throws InvalidZoneOffsetException
      */
     public function __construct($time = 'today', DateTimeZone $timezone = null, string $format = null) {
         if (null === $timezone) {
             $timezone = new DateTimeZone('UTC');
+        } elseif ('UTC' !== $timezone->getName()) {
+            throw new InvalidZoneOffsetException('Invalid zone offset: '.$timezone->getName().'. Should be 0');
         }
         if ('today' !== $time) {
-            $format = $format ?? $this->FORMAT;
-
-            $parsedDate = (object) date_parse_from_format($format, $time);
-            if ($parsedDate->error_count > 0) {
-                throw new InvalidDateFormatException('Invalid date format: '.$time.'. Should be '.$format);
-            }
+            $this->checkFormat($time, $format);
         }
         parent::__construct($time, $timezone);
-    }
-
-    public function __toString() {
-        return $this->format($this->FORMAT);
-    }
-
-    /**
-     * formatter used with json_encode.
-     */
-    public function jsonSerialize() {
-        return $this->__toString();
     }
 }
